requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==1.98.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
starlette==0.44.0
tqdm==4.67.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
